+++ plugins/modules/nomad_namespace.py
@@ -13,26 +13,26 @@ from ..module_utils.utils import del_none, is_subset
 
 def run_module():
     # define available arguments/parameters a user can pass to the module
-    module_args = dict(
-        state=dict(type="str", choices=["present", "absent"], default="present"),
-        url=dict(type="str", required=True, fallback=(env_fallback, ["NOMAD_ADDR"])),
-        validate_certs=dict(type="bool", default=True),
-        connection_timeout=dict(type="int", default=10),
-        management_token=dict(
-            type="str",
-            required=True,
-            no_log=True,
-            fallback=(env_fallback, ["NOMAD_TOKEN"]),
-        ),
-        name=dict(type="str", required=True),
-        description=dict(type="str"),
-        meta=dict(type="str"),
-    )
+    module_args = {
+        "state": {"type": "str", "choices": ["present", "absent"], "default": "present"},
+        "url": {"type": "str", "required": True, "fallback": (env_fallback, ["NOMAD_ADDR"])},
+        "validate_certs": {"type": "bool", "default": True},
+        "connection_timeout": {"type": "int", "default": 10},
+        "management_token": {
+            "type": "str",
+            "required": True,
+            "no_log": True,
+            "fallback": (env_fallback, ["NOMAD_TOKEN"]),
+        },
+        "name": {"type": "str", "required": True},
+        "description": {"type": "str"},
+        "meta": {"type": "str"},
+    }
 
     # seed the final result dict in the object. Default nothing changed ;)
-    result = dict(
-        changed=False,
-    )
+    result = {
+        "changed": False,
+    }
 
     # the AnsibleModule object
     module = AnsibleModule(argument_spec=module_args, supports_check_mode=False)
@@ -42,18 +42,17 @@ def run_module():
 
     existing_namespace = nomad.get_namespace(module.params.get("name"))
     desired_namespace = del_none(
-        dict(
-            Name=module.params.get("name"),
-            Description=module.params.get("description"),
-            Meta=module.params.get("meta"),
-            Capabilities=module.params.get("capabilities"),
-        )
+        {
+            "Name": module.params.get("name"),
+            "Description": module.params.get("description"),
+            "Meta": module.params.get("meta"),
+            "Capabilities": module.params.get("capabilities"),
+        }
     )
 
-    if module.params.get("state") == "absent":
-        if existing_namespace is not None:
-            nomad.delete_namespace(module.params.get("name"))
-            result["changed"] = True
+    if module.params.get("state") == "absent" and existing_namespace is not None:
+        nomad.delete_namespace(module.params.get("name"))
+        result["changed"] = True
 
     # decide if we should create/update a namespace
     create_namespace = False
